refactor(cli): issue_spec.rs を mod.rs + structured.rs + tests.rs に分割 (SPEC-1942)
cli/issue_spec.rs (1552 行) を 3 sub-file に分割し god-file を完全消滅。
SPEC-1942 SC-027 listed family scope 外の sub-module だが、 1500+ 行の
god-file はアーキテクチャ的負債なので一括解消。

新構成:
- cli/issue_spec/mod.rs (503 行): parse + run dispatch + SPEC_CREATE_HELP +
  SPEC_SECTION_NAME 定数。 SC-027 範囲内 (200-600)。
- cli/issue_spec/structured.rs (465 行): TextBlock + StructuredSpecInput +
  StructuredUserStory + parse_structured_spec_json + render_structured_spec
  + merge_structured_spec + split_structured_spec + render_*_section /
  normalize_*/ strip_* helpers。 構造化 JSON モデルと Markdown render を集約。
- cli/issue_spec/tests.rs (621 行、 cfg(test)): 全 test fn。 #[cfg(test)]
  mod tests; 経由で mod.rs から取り込み。

Visibility: structured.rs items は pub(super) として mod.rs から公開、
test 専用 import (#[cfg(test)] use structured::{...}) と production import
(use structured::{...}) を分離。 io_as_api_error は structured.rs に local
定義 (cli.rs の同名 fn とは独立)。

機能・契約・test 結果は不変 (lib 345 全 pass、 clippy 0 / fmt clean)。
god-file 完全消滅で cli/ ディレクトリ内 file 全てが <800 行範囲。
